[CL-2008] Extract donut chart
