+++ front/app/modules/commercial/analytics/admin/components/PostFeedbackCard/DonutChart/CenterLabel.tsx
